requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.1

# Hugging Face + LangChain
transformers>=4.40.0
huggingface-hub>=0.22.0
langchain-core>=0.2.20
langchain-huggingface>=0.1.0

# LangGraph
langgraph>=0.0.50

# Data validation
pydantic>=2.6.0

# Optional but recommended (tokenization stability)
sentencepiece>=0.1.99
